requirements.txt
flask
psycopg2-binary
python-dotenv
openai
openai-whisper
requests
ffmpeg-python
